# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SubmitAnalysisRequest {
  /** The URL of the website to analyse */
  url: string;
}

export type AnalysisJobStatus =
  (typeof AnalysisJobStatus)[keyof typeof AnalysisJobStatus];

export const AnalysisJobStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

export interface AnalysisJob {
  id: string;
  url: string;
  status: AnalysisJobStatus;
  createdAt: string;
  completedAt?: string | null;
  error?: string | null;
}

export type AnalysisSummaryStatus =
  (typeof AnalysisSummaryStatus)[keyof typeof AnalysisSummaryStatus];

export const AnalysisSummaryStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

export interface AnalysisSummary {
  id: string;
  url: string;
  status: AnalysisSummaryStatus;
  screenshotUrl?: string | null;
  overallScore?: number | null;
  createdAt: string;
  completedAt?: string | null;
}

export type FindingType = (typeof FindingType)[keyof typeof FindingType];

export const FindingType = {
  strength: "strength",
  weakness: "weakness",
} as const;

export interface Finding {
  id: string;
  type: FindingType;
  /** Dynamic category determined by AI based on page intent */
  category: string;
  title: string;
  evidence: string;
  suggestion: string;
  /** X position as a percentage of the screenshot width (0-100) */
  x: number;
  /** Y position as a percentage of the screenshot height (0-100) */
  y: number;
  /** Width as a percentage of the screenshot width */
  width: number;
  /** Height as a percentage of the screenshot height */
  height: number;
}

export interface ScoreCategory {
  name: string;
  score: number;
  description: string;
}

export type AnalysisStatus =
  (typeof AnalysisStatus)[keyof typeof AnalysisStatus];

export const AnalysisStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

export interface Analysis {
  id: string;
  url: string;
  status: AnalysisStatus;
  screenshotUrl?: string | null;
  screenshotWidth?: number | null;
  screenshotHeight?: number | null;
  overallScore?: number | null;
  findings: Finding[];
  scores: ScoreCategory[];
  summary?: string | null;
  /** The AI-detected purpose of the page (e.g. Lead Generation, SaaS / Software) */
  pageIntent?: string | null;
  createdAt: string;
  completedAt?: string | null;
  error?: string | null;
}

export type SeoFindingStatus =
  (typeof SeoFindingStatus)[keyof typeof SeoFindingStatus];

export const SeoFindingStatus = {
  good: "good",
  warning: "warning",
  error: "error",
} as const;

export interface SeoFinding {
  category: string;
  title: string;
  status: SeoFindingStatus;
  details: string;
  recommendation: string;
}

export interface SeoAnalysis {
  analysisId: string;
  overallScore: number;
  findings: SeoFinding[];
  summary: string;
}