# lib/api-zod/src/generated/types/analysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisStatus } from "./analysisStatus";
import type { Finding } from "./finding";
import type { ScoreCategory } from "./scoreCategory";

export interface Analysis {
  id: string;
  url: string;
  status: AnalysisStatus;
  screenshotUrl?: string | null;
  screenshotWidth?: number | null;
  screenshotHeight?: number | null;
  overallScore?: number | null;
  findings: Finding[];
  scores: ScoreCategory[];
  summary?: string | null;
  /** The AI-detected purpose of the page (e.g. Lead Generation, SaaS / Software) */
  pageIntent?: string | null;
  createdAt: Date;
  completedAt?: Date | null;
  error?: string | null;
}