# lib/api-zod/src/generated/types/analysisJob.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisJobStatus } from "./analysisJobStatus";

export interface AnalysisJob {
  id: string;
  url: string;
  status: AnalysisJobStatus;
  createdAt: Date;
  completedAt?: Date | null;
  error?: string | null;
}