# lib/api-zod/src/generated/types/analysisSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisSummaryStatus } from "./analysisSummaryStatus";

export interface AnalysisSummary {
  id: string;
  url: string;
  status: AnalysisSummaryStatus;
  screenshotUrl?: string | null;
  overallScore?: number | null;
  createdAt: Date;
  completedAt?: Date | null;
}