# lib/api-zod/src/generated/types/finding.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FindingType } from "./findingType";

export interface Finding {
  id: string;
  type: FindingType;
  /** Dynamic category determined by AI based on page intent */
  category: string;
  title: string;
  evidence: string;
  suggestion: string;
  /** X position as a percentage of the screenshot width (0-100) */
  x: number;
  /** Y position as a percentage of the screenshot height (0-100) */
  y: number;
  /** Width as a percentage of the screenshot width */
  width: number;
  /** Height as a percentage of the screenshot height */
  height: number;
}