# lib/api-zod/src/generated/types/seoAnalysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SeoFinding } from "./seoFinding";

export interface SeoAnalysis {
  analysisId: string;
  overallScore: number;
  findings: SeoFinding[];
  summary: string;
}