# lib/api-zod/src/generated/types/submitAnalysisRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SubmitAnalysisRequest {
  /** The URL of the website to analyse */
  url: string;
}